- updated database layer and added nosql db boltDB - updated Readme and postman collection - updated some error status codes
